user update & delete

+++ src/main/java/com/cards/cards/services/UserService.java
@@ -6,6 +6,10 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.method.P;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -13,6 +17,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.client.HttpClientErrorException.NotFound;
 
 import com.cards.cards.dao.DeleteUserDao;
 import com.cards.cards.dao.UserDao;
@@ -31,7 +36,7 @@ public class UserService implements UserDetailsService {
     private EmailRepository _emailRepository;
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public UserDao createUser(UserDao user) {
+    public ResponseEntity<String> createUser(UserDao user) {
 
         /*
          * {
@@ -44,17 +49,37 @@ public class UserService implements UserDetailsService {
          */
 
         UserModel newUser = new UserModel(user.getName(), user.getDate(), user.getPassword());
-        _userRepository.save(newUser);
-        _emailRepository.save(new EmailData(newUser, user.getEmail()));
-        return user;
+        if (_userRepository.findFirstByName(newUser.getName()) != null) {
+            return new ResponseEntity<String>("The user exists!", HttpStatus.BAD_REQUEST);
+        } else {
+            _userRepository.save(newUser);
+            _emailRepository.save(new EmailData(newUser, user.getEmail()));
+            return new ResponseEntity<String>("User has created.", HttpStatus.OK);
+        }
     }
 
-    @Transactional
-    public String deleteUser(DeleteUserDao user) {
+    @Transactional(propagation = Propagation.REQUIRED)
+    public ResponseEntity<String> deleteUser(DeleteUserDao user) {
         UserModel deleteUserId = _userRepository.findFirstByName(user.getName());
-        _emailRepository.deleteAll(_emailRepository.findByUserId(deleteUserId));
-        _userRepository.delete(deleteUserId);
-        return "User has deleted.";
+        if (deleteUserId != null) {
+            _emailRepository.deleteAll(_emailRepository.findByUserId(deleteUserId));
+            _userRepository.delete(deleteUserId);
+            return new ResponseEntity<String>("User has deleted.", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<String>("User not found!", HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @Transactional(propagation = Propagation.REQUIRED)
+    public ResponseEntity<String> updateUser(UserDao user) {
+        UserModel updateUser = _userRepository.findFirstByName(user.getName());
+        if (updateUser != null) {
+            _emailRepository.updateUser(updateUser, user.getEmail());
+            _userRepository.save(new UserModel(updateUser.getId(), user.getName(), user.getDate(), user.getPassword()));
+            return new ResponseEntity<String>("User has updated.", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<String>("User not found!", HttpStatus.NOT_FOUND);
+        }
     }
 
     @Override
